# swagger.json
{"openapi":"3.0.0","paths":{"/users":{"post":{"operationId":"UserController_register","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"UserController_getAllUsers","parameters":[],"responses":{"200":{"description":""}}}},"/users/{id}":{"patch":{"operationId":"UserController_partialUpdate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"UserController_deleteUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"get":{"operationId":"UserController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/books":{"post":{"operationId":"BookController_create","parameters":[],"responses":{"201":{"description":""}}},"get":{"operationId":"BookController_getAllBooks","parameters":[],"responses":{"200":{"description":""}}}},"/books/avaiable":{"get":{"operationId":"BookController_getAllAvaiableBooks","parameters":[],"responses":{"200":{"description":""}}}},"/books/emprestados":{"get":{"operationId":"BookController_getLendedBooks","parameters":[],"responses":{"200":{"description":""}}}},"/books/{id}":{"get":{"operationId":"BookController_getBook","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"BookController_deleteBook","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"patch":{"operationId":"BookController_partialUpdate","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/books/{id}/rating":{"get":{"operationId":"BookController_getAverageRating","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/review":{"post":{"operationId":"ReviewController_create","parameters":[],"responses":{"201":{"description":""}}}},"/review/books/{id}":{"get":{"operationId":"ReviewController_getAllReviews","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/review/{id}":{"get":{"operationId":"ReviewController_getReview","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}},"delete":{"operationId":"ReviewController_deleteReview","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}}}},"/auth":{"post":{"operationId":"AuthController_signIn","parameters":[],"responses":{"201":{"description":""}}}}},"info":{"title":"API Principal Bookify","description":"API das principais funções do Bookify","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{}}}